+++ src/reactviews/pages/ConnectionDialog/connectionStringPage.tsx
@@ -6,7 +6,10 @@
 import { useContext } from "react";
 import { ConnectionDialogContext } from "./connectionDialogStateProvider";
 import { FormField, useFormStyles } from "../../common/forms/form.component";
-import { IConnectionDialogProfile } from "../../../sharedInterfaces/connectionDialog";
+import {
+    ConnectionDialogWebviewState,
+    IConnectionDialogProfile,
+} from "../../../sharedInterfaces/connectionDialog";
 import { FormItemSpec } from "../../common/forms/form";
 import { ConnectButton } from "./components/connectButton.component";
 import { TextareaProps } from "@fluentui/react-components";
@@ -27,9 +30,10 @@ export const ConnectionStringPage = () => {
                 context={connectionDialogContext}
                 component={
                     connectionDialogContext.state.connectionComponents
-                        .components[
-                        "connectionString"
-                    ] as FormItemSpec<IConnectionDialogProfile>
+                        .components["connectionString"] as FormItemSpec<
+                        ConnectionDialogWebviewState,
+                        IConnectionDialogProfile
+                    >
                 }
                 idx={index}
                 props={{ orientation: "horizontal" }}
